Introduce FaDuotoneIconDirective and refactor FaIconComponent to use host directives
